+++ packages/backend-next/src/index.ts
@@ -30,11 +30,45 @@ import { searchModuleTechDocsCollator } from '@backstage/plugin-search-backend-m
 import { searchPlugin } from '@backstage/plugin-search-backend/alpha';
 import { techdocsPlugin } from '@backstage/plugin-techdocs-backend/alpha';
 import { todoPlugin } from '@backstage/plugin-todo-backend';
+import { entityFeedbackPlugin } from '@backstage/plugin-entity-feedback-backend';
 import { catalogModuleUnprocessedEntities } from '@backstage/plugin-catalog-backend-module-unprocessed';
+import { badgesPlugin } from '@backstage/plugin-badges-backend';
+import { azureDevOpsPlugin } from '@backstage/plugin-azure-devops-backend';
+import { linguistPlugin } from '@backstage/plugin-linguist-backend';
+import { devtoolsPlugin } from '@backstage/plugin-devtools-backend';
+import { TaskScheduleDefinition } from '@backstage/backend-tasks';
 
 const backend = createBackend();
 
 backend.add(appPlugin({ appPackageName: 'example-app' }));
+
+// Badges
+backend.add(badgesPlugin());
+
+// Azure DevOps
+backend.add(azureDevOpsPlugin());
+
+// DevTools
+backend.add(devtoolsPlugin());
+
+// Entity Feedback
+backend.add(entityFeedbackPlugin());
+
+// Linguist
+const linguistSchedule: TaskScheduleDefinition = {
+  frequency: { minutes: 2 },
+  timeout: { minutes: 15 },
+  initialDelay: { seconds: 15 },
+};
+
+backend.add(
+  linguistPlugin({
+    schedule: linguistSchedule,
+    age: { days: 30 },
+    batchSize: 2,
+    useSourceLocation: false,
+  }),
+);
 
 // Todo
 backend.add(todoPlugin());