+++ plugins/kubernetes-backend/src/service/KubernetesProxy.ts
@@ -93,14 +93,12 @@ export class KubernetesProxy {
   ): RequestHandler {
     const { permissionApi } = options;
     return async (req, res, next) => {
-      const token = getBearerTokenFromAuthorizationHeader(
-        req.header('authorization'),
-      );
-
       const authorizeResponse = await permissionApi.authorize(
         [{ permission: kubernetesProxyPermission }],
         {
-          token,
+          token: getBearerTokenFromAuthorizationHeader(
+            req.header('authorization'),
+          ),
         },
       );
       const auth = authorizeResponse[0];
@@ -110,15 +108,19 @@ export class KubernetesProxy {
         return;
       }
 
-      req.headers.authorization =
-        req.header(HEADER_KUBERNETES_AUTH) ??
-        `Bearer ${
-          (
-            await this.getClusterForRequest(req).then(cd =>
-              this.authTranslator.decorateClusterDetailsWithAuth(cd, {}),
-            )
-          ).serviceAccountToken
-        }`;
+      const authHeader = req.header(HEADER_KUBERNETES_AUTH);
+      if (authHeader) {
+        req.headers.authorization = authHeader;
+      } else {
+        const { serviceAccountToken } = await this.getClusterForRequest(
+          req,
+        ).then(cd =>
+          this.authTranslator.decorateClusterDetailsWithAuth(cd, {}),
+        );
+        if (serviceAccountToken) {
+          req.headers.authorization = `Bearer ${serviceAccountToken}`;
+        }
+      }
 
       const middleware = await this.getMiddleware(req);
       middleware(req, res, next);
@@ -135,12 +137,23 @@ export class KubernetesProxy {
       const logger = this.logger.child({ cluster: originalCluster.name });
       middleware = createProxyMiddleware({
         logProvider: () => logger,
+        ws: true,
         secure: !originalCluster.skipTLSVerify,
         changeOrigin: true,
+        pathRewrite: async (path, req) => {
+          // Re-evaluate the cluster on each request, in case it has changed
+          const cluster = await this.getClusterForRequest(req);
+          const url = new URL(cluster.url);
+          return path.replace(
+            new RegExp(`^${req.baseUrl}`),
+            url.pathname || '',
+          );
+        },
         router: async req => {
           // Re-evaluate the cluster on each request, in case it has changed
           const cluster = await this.getClusterForRequest(req);
           const url = new URL(cluster.url);
+
           return {
             protocol: url.protocol,
             host: url.hostname,
@@ -148,7 +161,6 @@ export class KubernetesProxy {
             ca: bufferFromFileOrString('', cluster.caData)?.toString(),
           };
         },
-        pathRewrite: { [`^${originalReq.baseUrl}`]: '' },
         onError: (error, req, res) => {
           const wrappedError = new ForwardedError(
             `Cluster '${originalCluster.name}' request error`,
@@ -173,7 +185,7 @@ export class KubernetesProxy {
   }
 
   private async getClusterForRequest(req: Request): Promise<ClusterDetails> {
-    const clusterName = req.header(HEADER_KUBERNETES_CLUSTER);
+    const clusterName = req.headers[HEADER_KUBERNETES_CLUSTER.toLowerCase()];
     const clusters = await this.clusterSupplier.getClusters();
 
     if (!clusters || clusters.length <= 0) {