+++ plugins/airbrake/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @backstage/plugin-airbrake
 
+## 0.3.20-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/theme@0.4.1-next.0
+  - @backstage/core-components@0.13.3-next.1
+  - @backstage/core-plugin-api@1.5.3-next.0
+  - @backstage/dev-utils@1.0.17-next.1
+  - @backstage/test-utils@1.4.1-next.1
+  - @backstage/plugin-catalog-react@1.7.1-next.1
+
+## 0.3.20-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.3-next.0
+  - @backstage/catalog-model@1.4.1-next.0
+  - @backstage/core-plugin-api@1.5.2
+  - @backstage/dev-utils@1.0.17-next.0
+  - @backstage/test-utils@1.4.1-next.0
+  - @backstage/theme@0.4.0
+  - @backstage/plugin-catalog-react@1.7.1-next.0
+
 ## 0.3.19
 
 ### Patch Changes