+++ plugins/app-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-app-backend
 
+## 0.3.47-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.1-next.0
+  - @backstage/backend-plugin-api@0.5.4-next.0
+  - @backstage/config@1.0.8
+  - @backstage/config-loader@1.3.2-next.0
+  - @backstage/types@1.1.0
+
 ## 0.3.46
 
 ### Patch Changes