+++ plugins/catalog-customized/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @internal/plugin-catalog-customized
 
+## 0.0.12-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog@1.11.3-next.1
+  - @backstage/plugin-catalog-react@1.7.1-next.1
+
+## 0.0.12-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog@1.11.3-next.0
+  - @backstage/plugin-catalog-react@1.7.1-next.0
+
 ## 0.0.11
 
 ### Patch Changes