+++ plugins/events-backend/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-events-backend
 
+## 0.2.8-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.1-next.0
+  - @backstage/backend-plugin-api@0.5.4-next.0
+  - @backstage/config@1.0.8
+  - @backstage/plugin-events-node@0.2.8-next.0
+
 ## 0.2.7
 
 ### Patch Changes