+++ plugins/lighthouse-backend/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-lighthouse-backend
 
+## 0.2.3-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.1-next.0
+  - @backstage/backend-tasks@0.5.4-next.0
+  - @backstage/catalog-client@1.4.3-next.0
+  - @backstage/catalog-model@1.4.1-next.0
+  - @backstage/config@1.0.8
+  - @backstage/types@1.1.0
+  - @backstage/plugin-lighthouse-common@0.1.2
+
 ## 0.2.2
 
 ### Patch Changes