+++ plugins/permission-react/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-permission-react
 
+## 0.4.14-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.5.3-next.0
+  - @backstage/config@1.0.8
+
+## 0.4.14-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.0.8
+  - @backstage/core-plugin-api@1.5.2
+  - @backstage/plugin-permission-common@0.7.7-next.0
+
 ## 0.4.13
 
 ### Patch Changes