+++ plugins/playlist/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @backstage/plugin-playlist
 
+## 0.1.12-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/theme@0.4.1-next.0
+  - @backstage/core-components@0.13.3-next.1
+  - @backstage/core-plugin-api@1.5.3-next.0
+  - @backstage/plugin-catalog-react@1.7.1-next.1
+  - @backstage/plugin-search-react@1.6.3-next.1
+  - @backstage/plugin-permission-react@0.4.14-next.1
+
+## 0.1.12-next.0
+
+### Patch Changes
+
+- 68531cc70e1f: Break some internal circular import chains
+- Updated dependencies
+  - @backstage/errors@1.2.1-next.0
+  - @backstage/core-components@0.13.3-next.0
+  - @backstage/catalog-model@1.4.1-next.0
+  - @backstage/core-plugin-api@1.5.2
+  - @backstage/theme@0.4.0
+  - @backstage/plugin-catalog-common@1.0.15-next.0
+  - @backstage/plugin-catalog-react@1.7.1-next.0
+  - @backstage/plugin-permission-common@0.7.7-next.0
+  - @backstage/plugin-permission-react@0.4.14-next.0
+  - @backstage/plugin-playlist-common@0.1.8-next.0
+  - @backstage/plugin-search-react@1.6.3-next.0
+
 ## 0.1.11
 
 ### Patch Changes