+++ plugins/sonarqube-backend/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-sonarqube-backend
 
+## 0.2.1-next.1
+
+### Patch Changes
+
+- fcaf4cfc70ee: Made default config optional
+- Updated dependencies
+  - @backstage/config@1.0.8
+
+## 0.2.1-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/errors@1.2.1-next.0
+  - @backstage/backend-common@0.19.1-next.0
+  - @backstage/config@1.0.8
+
 ## 0.2.0
 
 ### Minor Changes