+++ plugins/stack-overflow/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @backstage/plugin-stack-overflow
 
+## 0.1.18-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/theme@0.4.1-next.0
+  - @backstage/core-components@0.13.3-next.1
+  - @backstage/core-plugin-api@1.5.3-next.0
+  - @backstage/plugin-search-react@1.6.3-next.1
+  - @backstage/plugin-home-react@0.1.1-next.1
+  - @backstage/config@1.0.8
+
+## 0.1.18-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.3-next.0
+  - @backstage/config@1.0.8
+  - @backstage/core-plugin-api@1.5.2
+  - @backstage/theme@0.4.0
+  - @backstage/plugin-home-react@0.1.1-next.0
+  - @backstage/plugin-search-common@1.2.5-next.0
+  - @backstage/plugin-search-react@1.6.3-next.0
+
 ## 0.1.17
 
 ### Patch Changes