+++ plugins/stackstorm/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @backstage/plugin-stackstorm
 
+## 0.1.4-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/theme@0.4.1-next.0
+  - @backstage/core-components@0.13.3-next.1
+  - @backstage/core-plugin-api@1.5.3-next.0
+
+## 0.1.4-next.0
+
+### Patch Changes
+
+- 1745959daf98: Add props to StackstormHome for Header Customization
+- Updated dependencies
+  - @backstage/errors@1.2.1-next.0
+  - @backstage/core-components@0.13.3-next.0
+  - @backstage/core-plugin-api@1.5.2
+  - @backstage/theme@0.4.0
+
 ## 0.1.3
 
 ### Patch Changes