+++ plugins/user-settings-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-user-settings-backend
 
+## 0.1.11-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/errors@1.2.1-next.0
+  - @backstage/backend-common@0.19.1-next.0
+  - @backstage/backend-plugin-api@0.5.4-next.0
+  - @backstage/catalog-model@1.4.1-next.0
+  - @backstage/types@1.1.0
+  - @backstage/plugin-auth-node@0.2.16-next.0
+
 ## 0.1.10
 
 ### Patch Changes